Move corecli package to internal package so it isn't accidentally part of a public API
