+++ pkg/internal/corecli/corecli_stub.go
