+++ pkg/internal/corecli/extract.go
